Jetpack: Remove wp-windows8 gallery shortcode special case in WPCOM API post endpoint (#37243)
* Jetpack: Remove wp-windows8 gallery shortcode special case in WPCOM API post endpoint

This special case was originally added in r46288-wpcom back in 2012,
with a comment about "gallery shortcode overwrite for Windows Metro
app", and seems to have been largely unchanged since.

The special case seems to have been based on WordPress core's
`gallery_shortcode()` function at the time, with various pieces removed
as part of whatever special casing was needed. Unfortunately one of the
pieces removed was the `return $output;` at the end, making the
shortcode render to nothing rather than whatever output was intended.

As far as I've been able to tell, in 2015 Automattic released a new
"WordPress.com App for Windows" replacing that Metro app; looking
through archive.org, the Metro app itself seems to have disappeared from
Microsoft's app store sometime between [Feb 8] and Feb 23, 2014.

I don't find "wp-windows8" anywhere in the history of the code for the
[new app], and a quick check of the nginx logs in logstash didn't show any
instances of the "wp-windows8" user agent in the past 30 days.

Given all this, I think it's safe to remove this code. If someone
disagrees, we could still drop the `win8_gallery_shortcode` method in
favor of simply `__return_null`.

[Feb 8]: https://web.archive.org/web/20140208121528/http://apps.microsoft.com:80/windows/en-us/app/wordpress-com/a47ceb1d-b47c-48b0-ba1c-b81bb6ed2220
[new app]: https://github.com/Automattic/wp-desktop

* Update phan baseline