Nav: Handle plugins menu in redesigned wpcom nav
In Calypso the Plugins menu item always points to /plugins/:site
In wp-admin the Plugins menu always points to plugins.php

Plugins.php will include a button directing people to the WPCom
marketplace, and if necessary will also prompt them to upgrade.

This change ensures that the Plugins menu it always correctly
highlighted in the menu when on a Plugins page.