Nav: Highlight plugins menu in redesign
If the user can install plugins for this site then the Plugins menu item
links to the WPCom plugins marketplace. If the user can't install
plugins for this site and they can upgrade the sites plan then they
should be linked to plugins.php which shows a banner explaining so. If
this user ends up in Calypso (e.g. by using the Browse plugins link on
the banner), then the plugins menu will link to the wpcom marketplace.