Social: Fix the social icon colours on the admin page
The social icon colours were not being picked up on the admin page,
which turned out to be a problem with our Post CSS configuration.

In addition we were using --jp-gray-5 which wasn't defined by the
standard theme, but is the same value as --jp-gray. This has been a
problem before, so instead of swapping the property name, I've added
--jp-gray-5 as an alias.