phan: Add a plugin for handling "html template" files
An "html template" file is a PHP file that produces a fragment of HTML,
loaded by a method or function call somewhere else in the codebase.

Phan will normally evaluate the template as being in the global scope,
meaning it doesn't know about `$this` or other variables and will
complain about attempts to access them.

This plugin adds support for `@html-template` and `@html-template-var`
annotations to declare that a file is a template, what its context is,
and what variables are intended to be available.