cli: Remove test-only deps when building for mirrors
We added the idea of test-only deps in #37140, to allow Phan to be able
to analyze packages with circular dependencies while still being able to
determine a consistent build order.

Turns out that when building for the mirrors, composer gets confused if
there are depended-on packages existing in the output dir that aren't
listed in the composer.json `.repositories.[].options.versions` field.
To avoid this, let's have the for-mirrors build delete the test-only
dependencies from `require-dev` so Composer won't try to use them.