Carousel: be more defensive when checking for info from attachment
This should avoid errors like this one:

```
Warning: Attempt to read property "ID" on null in wp-content/plugins/jetpack/modules/carousel/jetpack-carousel.php on line 93
```

This also consolidates the code a bit, as a follow-up to 8d654cbb0cbf1f43cb38da96813fbe94de420666: we do not need to be fetching the full attachment post when it's already provided to us as a parameter of the method. We also do not need to check if attachment exists when we do it higher up in the method.

Internal discussion: p1716818804902439-slack-C05PV073SG3