constants: Fix type for `set_constant()` value
PHP's `define()` allows any of int|float|string|bool|array|null, so
let's go with that to match.

This fixes 57 Phan issues across the monorepo.

---

Just as I was about to create this PR, #36968 changed it to `scalar`.
Which fixed 54 of those 57 errors, but as it left out the possibility of
array and null 3 were still left.