tools: Check that `pnpm` works before piping output (#37541)
If the `engines` in `package.json` aren't satisfied, `pnpm run` (among
other `pnpm` commands, but not all of them) will print an error message
to standard output and exit with a non-zero status. When the first use
of `pnpm` in a script is doing like `pnpm jetpack dependencies | jq`,
the script winds up exiting with an unhelpful error like
```
jq: parse error: Invalid numeric literal at line 1, column 34
```

To avoid this, lets run some other `pnpm jetpack` command earlier in the
affected scripts. And to avoid useless or confusing output for the more
common case that things work properly, let's introduce a `jetpack noop`
command that does nothing beyond the normal startup stuff to use for
this.