Required Review Action: omit author from reviewer assignment (#37653)
This PR fixes a bug where the required-review action would attempt to
add the PR author as a reviewer if a matching rule was hit, however this
is no longer allowed by GitHub and would result in the error `Review
cannot be requested from pull request author.`. To avoid this, if
computed reviewers list includes the PR author, a message is logged and
they are filtered out of the list.