ci: Fix second fetch in monorepo autotagger
The monorepo autotagger fetches tags, determines if it needs to create
any tags, then if so it waits for any competing runs to complete and
then fetches tags again in case the competing run created some.

The problem happens if existing tags were _changed_, for example by the
PR is up-to-date workflow updatinng its own tags. In that case the
second fetch chokes with errors like
```
 ! [rejected]            pr-update-to-projects/js-packages/components -> pr-update-to-projects/js-packages/components  (would clobber existing tag)
```

To avoid this, pass `--force` to the second tag fetch.