cli: For `jetpack release version`, only init next cycle when asked to (#37718)
In #27641 we added code to init the next cycle to `jetpack release
version`. It's not clear to me now whether that fixed one use case while
breaking another, or whether the 'another' was added later.

Either way, we only want to start the next cycle in some cases. So add
an `--init-next-cycle` option to use for those cases.