renovate: Fix warnings from weird wpcomsh deps
For the five `type: vcs` deps, Renovate needs a `name` key in each one
so it knows the package name (versus Composer itself that fetches the
composer.json).

The "wordpress/classic-editor-plugin" fake package is a bit more
difficult. We have to disable the 'packagist' entry entirely. But then
we can make a customManager and customDatasource for it. 😀