eslint-config-target-es: Support flat config (eslint v9) (#37855)
This adds new entry points for the flat configs corresponding to the
existing eslintrc entry points, and for good measure adds explicit entry
points for the eslintrc versions as well. When we next do a major
version bump, we might swap the "default" entry points from eslintrc to
flat config.