cli: Fix rsync of unpublished plugins to plugins dir
When running `jetpack rsync xxx host:/path/to/wp-content/plugins`, the
command is supposed to look up the plugin slug so it can rsync to the
correct subdirectory. This works for published plugins (that have
`.extra.wp-plugin-slug` in composer.json), but for unpublished plugins
it winds up rsyncing to `/path/to/wp-content/plugins/undefined/`.

The fix is twofold:

* If `.extra.wp-plugin-slug` is not set, check `.extra.beta-plugin-slug`.
* If both fail to give us a slug, raise an error.