eslint: Remove `root: true` from `tools/e2e-commons/.eslintrc.cjs`
This is something of a proof of concept for some of the cleanup work
suggested in pf4qpu-rJ-p2, where we'd like to apply the standard
monorepo ruleset with targeted additions and overrides instead of using
`root: true` to start fresh more often.

In this case it turned out to be pretty straightforward, just tedious
since most of the things needing manual fixing were just writing missing
jsdoc comments.

Other stuff:
* Auto-fix `import` ordering.
* `loadIgnorePatterns` is no longer needed in `**/tests/e2e/.eslintrc.js`
  since we're no longer doing `root: true`.
* Suppress `testing-library/prefer-screen-queries` in e2e dirs,
  it false-positives on various Playwright `page` methods.
* Boost: Add a `tsconfig.eslint.json` to "include" the tests in a
  project while linting.
* Suppress 'no-process-exit' in a few tools/e2e-commons/bin/ scripts.
* A few miscellaneous refactors: no-case-declarations, a
  lodash/import-scope, a no-useless-escape, a no-constant-condition, and
  a few functions that only sometimes returned (unused) values.