beta: Support testing mu-plugins
Now that we have wpcomsh in the monorepo, people want to test it using
Jetpack Beta Tester. For that to work, we have to teach it how to handle
mu-plugins.

It's more than just pointing it at `WPMU_PLUGIN_DIR`, since WordPress
loads mu-plugins differently than it does normal plugins and has no
built-in concept of deactivating them. Fortunately we can take advantage
of how WordPress requires a "loader" file to load a mu-plugin to give us
a way to "activate" either the stable or dev version on demand.

The diff here is a bit bigger than it would otherwise be because we have
to replace a bunch of uses of `WP_PLUGIN_DIR` and `is_plugin_active()`
outside of upgrading code with stuff that can handle mu-plugins too.