Abandon google-fonts-provider package
The google-fonts-provider package made use of APIs that were provided by
the Gutenberg plugin versions v12.8.0 to v16.3.0 and never made it into
WordPress Core.

Meanwhile, the google-fonts module in the Jetpack plugin has been
updated to use an API that was included in WordPress Core 6.4, which is
the minimum version we support.

So let's abandon the package, and remove the back-compat code in the
Jetpack plugin that used it.