Blaze: bail earlier when there are connection issues
When a site or user is not properly connected to WordPress.com, or when there are sync issues on the site, we can estimate that the site is not eligible to Blaze, since they won't be able to create Blaze campaigns when those things are not working properly.

Let's consequently bail earlier, consider Blaze as not ready to be initialized. That will save us a lot of failed remote requests later.

Related discussion: p1719416484308139-slack-CBG1CP4EN