Improve performance of `changelogger-validate-all.sh`
`changelogger version` is unnecessarily calling `git` to fetch the
timestamp and PR number for all the changelog entries, when all it's
really using is the significance. Skip that.

Then, while validating each individual project is reasonably fast, when
there's 117 of them it adds up. We can improve things somewhat by
processing projects in parallel, matching the number of processors in
the system.

On my laptop, the first takes the time from ~16.8s to ~10.7s, and the
second (8 cores) takes it to ~2.3s.