Publicize: Ensure that connections are only retrieved once per request.
A race condition where the transient, used to cache the connections
information, gets set in a second request, while another is in flight,
causes the `set_transient` call to fail.

As we relied on the transient, this meant that each time we tried to
retrieve the current connection information, we would make a request to
the API, and then fail to set the transient again.

This change solves that problem by storing the information in memory and
also removing the transient if the call to `set_transient` fails.

There is an outstanding task where we need to work out why the
Connections_Post_Field causes so many calls to `get_all_connections`.