tools: Fix colors in scripts (#38142)
GitHub seems to have changed something since we wrote a lot of these.
The `-t` test seems to no longer be true, so nothing outputs colors.
Check for `GITHUB_ACTIONS` to be set to `true` and assume `-t` would be
true in that case (hopefully we're not doing too much piping of script
outputs).

Also, tests indicate that, despite not setting TERM or anything like
that, Actions supports 24-bit color output. So check for
`GITHUB_ACTIONS` in a few places where we use 256-color or 24-bit-color.

Also at least one person has the "bold is bright" setting in their
terminal off, so setting "bold-black" for debug text isn't working for
them. If `tput dim` returns something non-empty, use that, otherwise
hope that 90 is supported.

On the plus side, Actions output seems to do ok now with `90` so no need
to override `debug` in various scripts anymore.