Remove abandoned packages from the monorepo
These are packages/abtest and packages/google-fonts-provider.

Doing this will save some of our tooling having to check them, build
them, or run tests on them.

The old mirror repos will remain, should we decide we need to update
them in some manner despite their being abandoned. If we decide to
un-abandon them, the relevant parts of this PR can be reverted.