jetpack: Fix undefined array warnings in Donations block
If the donations block sets `oneTimeDonation`, `monthlyDonation`, and/or
`annualDonation` without including a `planId` field, ensure the field is
set by including the value used in the block.json in the defaults.

If the `currency` attribute is not set or is invalid, and
`showCustomAmount` is true, fall back to the generic currency symbol `¤`
like `Jetpack_Currencies::format_price()` already does and default to
¤100.

Also, `Jetpack_Currencies::format_price()` should honor its `$symbol`
argument when doing that fallback.