Odd there's a test for no-symbol no-currency producing the symbol
I don't see anything in #17381 that discusses this behavior. It's
possible it was an accidental bug that wasn't noticed because of the
undefined array key access producing null which produces the empty
string, and that the test was a copy-paste of the previous test that
didn't notice `false` being passed.

I see no other uses of `Jetpack_Currencies` with `$symbol = false` in
the monorepo or opengrok or wpdirectory, so this should be safe to
fix.