jetpack-mu-wpcom: Stabilize build (#38284)
PR #38093 added downloading of translation data into the build for the
package, and decided to store the current date in each file meaning that
every file changes for every single build. This clutters up the history
of the mirror repo, making it hard to see what the actual changes are
since every trunk commit has these useless date changes.

We can improve the situation somewhat by saving the HTTP Last-Modified
date into the files instead of the current date, so they should only be
updated when the translations actually change. We'll still get updates
attributed to whichever unrelated trunk commit happened to be the next
after a change on the translation server, but at least we won't get it
for every single commit.

Even better, IMO, would be if the translations could be downloaded to
sites as needed like WordPress does with translations for plugins
installed from wporg. But that would probably require a lot more
refactoring.