ci: Check for missing non-dev deps in plugins
Our build process by default ignores any files in `.gitignore`. If you
want to include such files, you have to set `production-include` in
`.gitattributes`.

For plugins, we want non-dev Composer deps to be included in the zip.
For the most part we do this by individually "production-including" the
appropriate dirs, mainly so we can keep an eye on if anything happens to
suddenly bring in an unexpected new dependency.

This adds a linting check to flag when a non-dev dependency isn't being
production-included, and fixes some missing entries.