jetpack: Use better cipher in WPCOM Connection test
RC4 is long obsolete, and likely isn't even available on modern hosts.

Choose AES-256-CTR or AES-256-CBC if available and use that instead. If
neither is available, don't even bother with RC4 anymore.

Note this needs D155629-code on the server, or the test in the Jetpack
Debugger will show as failing.