repo-gardening: Remove guessing of release dates
We added this guessing back when Jetpack was the only plugin, and we
still did a week-long testing week. That just got carried forward as we
added additional plugins and so on.

For Jetpack we should be safe relying on the milestone, it's on the
release checklist to create the new milestone. For most of the other
plugins, if there's no milestone then we may as well just say there's no
scheduled release. The exceptions to that are wpcomsh and
mu-wpcom-plugin, which get custom messaging if no milestone exists.