lazy-images: Mark package as abandoned (#38487)
WordPress Core has been outputting the necessary attributes for
browser-native lazy image loading for some time now, and it conflicts
with the new Interactivity API added by Gutenberg in 16.6. We've removed
use of it from both Jetpack and Boost. Let's finish the job by marking
it as deprecated, and then we can do a final release and remove it from
the monorepo.