Remove abandoned lazy-images package from the monorepo (#38513)
Doing this will save some of our tooling having to check it, build it,
or run tests on it.

The old mirror repo will remain, should we decide we need to update it
in some manner despite it being abandoned. If we decide to un-abandon
it, this PR can be reverted.