jetpack-mu-wpcom: Sync newspack blocks on build, not install
If you do it on install, everyone everywhere in the monorepo has to have
this downloaded and installed even if they're not working on mu-wpcom at
all.

Instead, save the syncing for the actual build, which seems to be the
only place it's needed. If it had been needed for testing too, we could
have added it to the `test-php` script or the like too.