OG Tags / Fediverse: check for active integration instead of class
Let's not check if the `Activitypub\Integration\Opengraph` class exists:

- This can be problematic in an autoloaded environment.
- The ActivityPub plugin actually offers an option allowing you to not require the class:
https://github.com/Automattic/wordpress-activitypub/blob/0bbd715d05fc8b2c399529872715605cffa613b1/integration/load.php#L41-L51

Unchecking that option actually causes a Fatal Error on the site, because of https://github.com/Automattic/wordpress-activitypub/blob/0bbd715d05fc8b2c399529872715605cffa613b1/activitypub.php#L126-L129

We now fully rely on the option instead. This has the inconvenient that if one deactivates and deletes the ActivityPub plugin, the option may stick around. This may be fixed if this issue gets fixed:
https://github.com/Automattic/wordpress-activitypub/issues/855