Add a bunch of `storybook` deps everywhere
Whey Storybook decided to do storybookjs/storybook#27039, they seem to
have neglected the use case where monorepos might only need a few
add-ons for stories in various packages, with all of the storybook stuff
in a separate dummy package. So now we need to bring in all of storybook
as a dev-dep in every package that has stories.