Block Editor: consider all blocks to be available on WoA
Do not attempt to check if a block is available on WoA. We consider that they all are, out of the box. It simplifies a logic, improves performance, and avoids relying on the "available" part of the plan information stored locally in the "jetpack_active_plan" option. This will allow us to remove that part from the local option, since it isn't used elsewhere.

Related conversation: p55Cj4-3q6-p2#comment-3642