Plan information: truncate data stored in local option
Until now, the "jetpack_active_plan" option included everything that was retrieved from the API. That made for a really big option, with a lot of data that we do not actually use in the codebase today (namely, the "available" part of the feature list).

With this commit, we remove that array from the data we store in the local option.

Related conversation: p55Cj4-3q6-p2#comment-3642