cli: Update version numbers pre-build
To stop updating version numbers to alpha versions in trunk while
keeping them in Packagist and such, we need to have the
build-for-mirrors update them, i.e. calling `tools/project-version.sh`
after we update the changelog, same as where we call
`tools/replace-next-version-tag.sh` during the build.

Since the updated version numbers affect what's stored in the
autoloader, we then have to move all this code up to the top of the
build process so it updates the version numbers before composer runs to
generate the autoloader.