Jetpack: lossless image optimization, part 3 (#38750)
Lossless image optimization of Jetpack images.

This is a continuation of #38398 and #38573. Please see #38398 for more context and details.

This finishes this series of commits. All images except for projects/js-packages/connection/components/disconnect-dialog/images/disconnect-thanks.jpg have been optimized. See @anomiex's debugging of the problem in https://github.com/Automattic/jetpack/pull/38750#pullrequestreview-2245838341 for more info on why that image was skipped.