eslint: Enable `array-callback-return` sniff (#39042)
By far the most common failure for this sniff was use of `.map()` as
`.forEach()`. These I turned into `for ( const x of y ) {}` loops, which
is even more straightforward than `.forEach()`.

Most of the rest were implicit returning of `undefined`, which got an
explicit return added.

A few cases were able to be simplified further, e.g. a loop that only
exists to set a boolean if anything matches can be replaced with
`.some()`.

Fixing this found one bug, in the now-unused wpcom-commit-reminder task
in repo-gardening.