tools: Improve bash `proceed_p` reading (#39038)
The function attempts to read any pending input before displaying the
prompt, so that any stray keypresses while the script was working won't
be interpreted as input to the prompt.

Turns out that the `-t 0` solution only works for full lines, if someone
typed text without hitting enter it wouldn't be caught and so would be
interpreted as input to the prompt.

We can fix this by using `-t 0.1` instead. Along with `-n` that'll read
partial lines too.