tools: Improve speed of `tools/changelogger-release.sh`
The script has gotten a bit slow, since it was running
`check-intra-monorepo-deps.sh` individually for each project to pass
either `-u` or `-U` depending on whether the project was included in the
release or not.

We can speed it up by running `check-intra-monorepo-deps.sh -U` just
once, and then looking at the git diff to see whether any changelog
entries seem to be needed.

Also, we can then drop the `PACKAGE_VERSIONS_CACHE` thing from
`check-intra-monorepo-deps.sh` that was done in a prior speedup attempt.

And finally, skip trying to handle the "monorepo" pseudo-package, it
won't work and prevents releases of projects depended on by the monorepo
root.