Replace `E_USER_ERROR` for PHP 8.4
It seems they've decided to deprecate using `E_USER_ERROR` in favor of
either dying or exceptions. The PHPCompatibility dev-branch run has
added a sniff for this, so we may as well fix it.

One instance, in a test, is downgraded to `E_USER_WARNING`. The other is
changed to a throw.