eslint: Enable `@wordpress/no-unused-vars-before-return` rule (#39127)
This rule tries to find cases where a variable is set to the result of a
function call, but then an early return happens before that value is
used. The idea being that we can likely save work by not calling a bunch
of functions unnecessarily.

Watch out for functions that have side effects, though. It may be that
those need to be called even if the return value doesn't seem to be
used. React hooks particularly fall into this category; when our
`jetpack-js-tools/eslintrc/react` config is extended any functions
beginning with "use" are ignored as such, but sometimes it seems we're
forgetting to extend that.