cli: Ask to add changelogs for indirectly affected plugins
A change to, for example, `packages/backup` might also be of interest
for the `plugins/backup` changelog. But as things stand, if nothing in
the PR touches `plugins/backup` then no changelog entry will be created
for it unless the PR author thinks to add one manually.

With this PR, `jetpack changelog add` will now ask whether such
indirectly-affected plugins should have a changelog added so we have a
better chance of entries being created.