eslint: Enable `react/jsx-key` rule
This rule requires a `key` prop be specified on elements that it thinks
probably need one, namely elements inside arrays or returned from
iterators for `.map()` and the like.

In practice most of the cases fixed here are probably pointless, since
the arrays are never going to be sorted or modified at runtime. But it's
easier to add a key than to write a comment justifing ignoring it in
each case.