jetpack: Render Slideshow block correctly in iframed editor
When the block editor uses an iframe, our monolithic `editor.js` script
is only loaded and run at the top level, not inside the iframe. Thus,
any dynamic `import()` also only runs at the top level, and in
particular the loaded CSS does not affect the content inside the iframe.

The real solution to this would be to have the content-related code in
block.json "script" where it gets loaded everywhere (frontend, editor
top-level, and iframe) (unfortunately there's no option that runs only
on frontend+iframe, or only iframe) and run the code doing the
`import()` inside the iframe, so it adds the CSS inside the iframe. But
we can't really do that without a ton of refactoring that's largely
blocked on Gutenberg implementing lazy block loading.

So, in the mean time, let's configure mini-css-extract-plugin to add any
imported CSS into the iframe (if one exists) in addition to the top
level. That won't work for JS (e.g. the Map block, which has its own
hack), but that's ok for now.