docker: Re-add xz-utils
Previously this was installed as a dependency of mailcap, which was a
dependency of mime-support, which was a dependency of various things.
But apparently things depending on mime-support are transitioning to
media-types which doesn't have that dependency.

We need xz-utils for extracting `build.tar.xz` in CI, so make sure it's
installed.