Make sitemap storing more efficient but not querying the full post content when storing the sitemap
This update tries to remediate a problem where the persistent object cache gets huge
with data that we will not use in the end.
We just need to get the id of the current sitemap post and if it exists, we update it.

Before, we were using get_post and this polluted the cache running into OOM situations