wpcomsh: Move `load_muplugin_textdomain` call to `after_setup_theme` hook
If `load_muplugin_textdomain` is called too early, it can cause problems
because it gets called before the user is determined. Core added a
doing-it-wrong check to the function that requires it be during or after
the `after_setup_theme` hook, which comes slightly after
`plugins_loaded` that we had been using.