Jetpack outdated WP notices: avoid fatals on outdated WP versions
Do not use the wp_admin_notice function to display that wp-admin notice. The notice is meant to be displayed on sites running an outdated version of WordPress, and it's possible that this version is 6.3-, which does not include the wp_admin_notice function.

Internal reference: p1728598216425289-slack-C03M3KD520Zd
Related PR, where wp_admin_notice was introduced: #37051