Jetpack: Stats: Fix csv trailing newline handling
When #39665 switched from incorrectly using `str_getcsv( $csv, "\n" )`
to `explode()`, a minor difference between the two was unnoticed: the
former does not produce an empty element at the end of the result on a
trailing newline, while the latter does.

Add an `rtrim` on the data to ensure there isn't a trailing newline.