changlogger: When amending, do not preserve default link
When doing `changelogger write --amend`, we normally preserve the
prologue, epilogue, and link unless these are overridden on the command
line.

An edge case is when you're amending version 1.2.3-beta to turn it into
1.2.3 release, and the link is the default link like
`https://example.com/diff/1.2.2...1.2.3-beta`. In this situation we
don't want to keep that link, we want to replace it with the new default
`https://example.com/diff/1.2.2...1.2.3`.