Improve image gallery accessibility
Accessibility has been improved by reducing which images receive a tab
index and an aria role of button.

This has been done by moving it out of PHP, where it applied the two
attributes to all images in the post, to setting it in javascript,
alongside where the interactivity is being setup.

An aria-label has been added for good measure.