ci: Initial code to process coverage data
To be able to inform the author about the coverage on their PR, we'll
need to extract a summary of coverage changes from the raw data we're
collecting. This is a first step towards that, combining the raw
coverage data into combined files and producing a summary of covered
lines in each file.

Also, since it annoyed me, I added `WordPress.WP.GlobalVariablesOverride`
to our `Jetpack-NoWP` ruleset so non-WordPress code doesn't get
useless complaints about common variable names like `$path`.