Migrate to WordPress.com: remove the plugin
The plugin is now developed by a partner, and no longer by a12s. The version currently available in the Jetpack monorepo no longer matches the version available in the WordPress.org plugin directory.

This commit removes all contents of the plugin in the monorepo, and only leaves indicators that this specific version (and its mirror repo and Packagist package) are no longer maintained, and archived.