Blocks: do not display block placeholder to non-admins
We currently display a placeholder in the block editor when a block cannot be used because its associated module is not active. The placeholder offers a button inviting folks to activate the module.

Since only admins can activate modules, the placeholder is useless to non-admins. Let's consequently skip registering the blocks for non-admins when the associated module is not active.