Open Graph: Always output tags if they haven't been.
Currently we only output the Open Graph tags if the sharing or Publicize
modules are enabled and there are no known conflicting plugins active.

This removes the module check, but adds a check where we capture the
output of wp_head, and then see if any of the tags have been output
already. If not then we render them.