Remove support links from several blocks
The group placeholder shows a choice of group blocks, hovering over them
shows [object Object]. This is because this code transforms them into
ReactElements (which isn't officially supported) and they get rendered
as strings.

Discussion is ongoing in Gutenberg about the best way to display the
support links. For now, to avoid the [object Object] problem, this pr
removes the links from the blocks that trigger the problem.