Repo Gardening: add Type labels to PRs based off changelog entries
Follow-up to #40428.

Let's try to automate the addition of "Type" labels to each PR.

When PR includes changelog files, and when that PR does not already has a [Type] label, let's scan the contents of the changelog files, pick the first "Type" definition we can find, and match that to one of our existing Type labels.